+++ config/config.php
@@ -11,18 +11,7 @@
  * @copyright Frank Hoppe 2014
  */
 
-/**
- * Backend-Bereich DSB anlegen, wenn noch nicht vorhanden
- */
-if(!$GLOBALS['BE_MOD']['dsb']) 
-{
-	$dsb = array(
-		'dsb' => array()
-	);
-	array_insert($GLOBALS['BE_MOD'], 0, $dsb);
-}
-
-$GLOBALS['BE_MOD']['dsb']['volunteeringlist'] = array
+$GLOBALS['BE_MOD']['content']['volunteeringlist'] = array
 (
 	'tables'         => array('tl_volunteeringlist', 'tl_volunteeringlist_items'),
 	'icon'           => 'system/modules/volunteeringlist/assets/images/icon.png',
@@ -35,11 +24,11 @@ $GLOBALS['BE_MOD']['dsb']['volunteeringlist'] = array
  */
 $GLOBALS['TL_CTE']['schach']['volunteeringlist'] = 'Samson\Volunteeringlist\Volunteeringlist';
 
-// Konfiguration für ProSearch
-$GLOBALS['PS_SEARCHABLE_MODULES']['volunteeringlist'] = array(
-	'icon'           => 'system/modules/volunteeringlist/assets/images/icon.png',
-	'title'          => array('title','name'),
-	'searchIn'       => array('title','name', 'info'),
-	'tables'         => array('tl_volunteeringlist', 'tl_volunteeringlist_items'),
-	'shortcut'       => 'vlist'
-);
+/**
+ * -------------------------------------------------------------------------
+ * Voreinstellungen
+ * -------------------------------------------------------------------------
+ */
+
+$GLOBALS['TL_CONFIG']['volunteeringlist_picWidth'] = 60;
+$GLOBALS['TL_CONFIG']['volunteeringlist_picHeight'] = 80;