+++ languages/de/tl_prosearch_data.php
@@ -1,3 +1,0 @@
-<?php
-
-$GLOBALS['TL_LANG']['tl_prosearch_data']['shortcut']['vlist'] = "Funktionärslisten";